backend/requirements.txt
# Backend Requirements for Supplier Clustering System

# Database
pymongo==4.6.0

# Web Framework
flask==3.0.0
flask-cors==4.0.0

# Authentication
PyJWT==2.8.0
bcrypt==4.1.2

# Data Processing (if needed in backend)
pandas==2.1.3
numpy==1.26.2

# Machine Learning (for clustering analysis)
scikit-learn==1.3.2
scipy==1.11.4
